refactor: Move hardhat related files to hardhat/ - This frees up the src/ directory to be able add typescript files related to the contracts
