+++ hardhat/index.ts
@@ -1,0 +1,5 @@
+import Task from './task';
+import Verifier from './evm/verifier';
+import { TaskRunOptions, Network } from './types';
+
+export { Task, TaskRunOptions, Verifier, Network }
