+++ tasks/20220816-deploy-lock/input.ts
@@ -1,4 +1,4 @@
-import { Network } from '../../src/types'
+import { Network } from '../../hardhat'
 
 export type DeploymentInputs = {
   unlockTime: number
