+++ tasks/20xxxxxx-template/input.ts
@@ -1,4 +1,4 @@
-import { Network } from '../../src/types'
+import { Network } from '../../hardhat'
 
 export type DeploymentInputs = {
   admin: string
